@@ -15,29 +15,32 @@ Gem::Specification.new do |s|
   s.description = "Set of addons to help with the activeadmin ui"
   s.license     = "MIT"
 
-  s.files = Dir["{app,config,db,lib,vendor}/**/*", "MIT-LICENSE", "Rakefile", "README.rdoc"]
+  s.files = Dir["{app,config,db,lib,vendor/assets}/**/*", "MIT-LICENSE", "Rakefile", "README.rdoc"]
 
   s.add_dependency "active_material"
   s.add_dependency "railties"
-  s.add_dependency "require_all", "~> 1.5"
-  s.add_dependency "sass"
-  s.add_dependency "select2-rails-latest", "~> 4.0"
+  s.add_dependency "require_all"
+  s.add_dependency "sassc"
+  s.add_dependency "sassc-rails"
   s.add_dependency "xdan-datetimepicker-rails", "~> 2.5.1"
 
   s.add_development_dependency "aasm"
-  s.add_development_dependency "enumerize", "~> 2.0"
-  s.add_development_dependency "paperclip"
-  s.add_development_dependency "rails", "~> 4.2"
-  s.add_development_dependency "sass-rails"
-  s.add_development_dependency "sqlite3"
-
   s.add_development_dependency "capybara-selenium"
-  s.add_development_dependency "chromedriver-helper"
   s.add_development_dependency "database_cleaner"
+  s.add_development_dependency "enumerize", "~> 2.0"
   s.add_development_dependency "factory_bot_rails"
   s.add_development_dependency "guard"
   s.add_development_dependency "guard-rspec"
+  s.add_development_dependency "image_processing"
+  s.add_development_dependency "paperclip"
   s.add_development_dependency "pry-rails"
+  s.add_development_dependency "puma"
+  s.add_development_dependency "rails", "~> 6.1.6"
   s.add_development_dependency "rspec-rails"
+  s.add_development_dependency "rspec_junit_formatter"
+  s.add_development_dependency "rubocop", "~> 0.65.0"
   s.add_development_dependency "shoulda-matchers"
+  s.add_development_dependency "shrine", "~> 3.0"
+  s.add_development_dependency "sqlite3"
+  s.add_development_dependency "webdrivers"
 end
