@@ -1,4 +1,4 @@
-class AddCategoryIdToInvoices < ActiveRecord::Migration
+class AddCategoryIdToInvoices < ActiveRecord::Migration[4.2]
   def change
     add_reference :invoices, :category, index: true
   end
