@@ -1,4 +1,4 @@
-class CreateRegions < ActiveRecord::Migration
+class CreateRegions < ActiveRecord::Migration[4.2]
   def change
     create_table :regions do |t|
       t.string :name
