@@ -1,4 +1,4 @@
-class AddCityToInvoices < ActiveRecord::Migration
+class AddCityToInvoices < ActiveRecord::Migration[4.2]
   def change
     add_reference :invoices, :city, index: true
   end
