@@ -1,4 +1,4 @@
-class AddInformationToRegions < ActiveRecord::Migration
+class AddInformationToRegions < ActiveRecord::Migration[4.2]
   def change
     add_column :regions, :information, :text
   end
