@@ -1,4 +1,4 @@
-class AddPositionToInvoices < ActiveRecord::Migration
+class AddPositionToInvoices < ActiveRecord::Migration[4.2]
   def change
     add_column :invoices, :position, :integer
   end
