@@ -1,4 +1,4 @@
-class AddColorAndAmountToInvoices < ActiveRecord::Migration
+class AddColorAndAmountToInvoices < ActiveRecord::Migration[4.2]
   def change
     add_column :invoices, :color, :string
     add_column :invoices, :amount, :float
