@@ -1,4 +1,4 @@
-class CreateInvoicesAndItems < ActiveRecord::Migration
+class CreateInvoicesAndItems < ActiveRecord::Migration[4.2]
   def change
     create_table :invoices_items, id: false do |t|
       t.belongs_to :invoice, index: true
