@@ -1,4 +1,4 @@
-class AddClientIdToInvoices < ActiveRecord::Migration
+class AddClientIdToInvoices < ActiveRecord::Migration[4.2]
   def change
     add_column :invoices, :client_id, :integer
   end
