@@ -1,4 +1,4 @@
-class CreateCars < ActiveRecord::Migration
+class CreateCars < ActiveRecord::Migration[4.2]
   def change
     create_table :cars do |t|
       t.string :name
