@@ -1,4 +1,4 @@
-class CreateManufacturers < ActiveRecord::Migration
+class CreateManufacturers < ActiveRecord::Migration[4.2]
   def change
     create_table :manufacturers do |t|
       t.string :name
