@@ -1,4 +1,4 @@
-class AddActiveToInvoices < ActiveRecord::Migration
+class AddActiveToInvoices < ActiveRecord::Migration[4.2]
   def change
     add_column :invoices, :active, :boolean, default: true
   end
