@@ -13,6 +13,7 @@ module ActiveAdminAddons
       toggle_classes = 'toggle-bool-switch'
       toggle_classes += ' on' if data
       toggle_classes += ' notify-success' if options[:success_message]
+      return unless enabled_controller_action?(:update)
 
       context.span(
         '',
@@ -22,7 +23,7 @@ module ActiveAdminAddons
         'data-object_id' => model.id,
         'data-field' => attribute,
         'data-value' => data,
-        'data-url' => context.auto_url_for(model),
+        'data-url' => resource_url,
         'data-success_message' => options[:success_message]
       )
     end
