@@ -1,17 +1,8 @@
 module ActiveAdminAddons
   module Rails
     class Engine < ::Rails::Engine
-      require "select2-rails"
-      require "sass"
-      begin
-        require 'sassc-rails'
-      rescue LoadError
-        begin
-          require 'sass-rails'
-        rescue LoadError
-          raise "Couldn't find 'sass-rails' or 'sassc-rails' gems in your application."
-        end
-      end
+      require 'sassc'
+      require 'sassc-rails'
       require "xdan-datetimepicker-rails"
       require "require_all"
       require "active_material"
