@@ -5,8 +5,18 @@ module ::Formtastic
 
       def default_input_type(method, options = {})
         input_type = original_default_input_type(method, options)
-        input_type = :datepicker if input_type == :date_select
+
+        if input_type == :date_select && active_admin_context?
+          input_type = :datepicker
+        end
+
         input_type
+      end
+
+      def active_admin_context?
+        options[:builder] == ::ActiveAdmin::FormBuilder
+      rescue NameError
+        false
       end
     end
   end
