@@ -1,3 +1,3 @@
 module ActiveadminAddons
-  VERSION = "1.6.0"
+  VERSION = "1.9.0"
 end
