@@ -1,19 +1,15 @@
 ActiveAdmin.register Invoice do
-  permit_params :legal_date, :number, :paid, :state, :attachment, :photo, :category_id,
-                :city_id, :amount, :color, :updated_at, :active, item_ids: [], other_item_ids: []
+  permit_params :legal_date, :number, :paid, :state, :attachment, :photo, :category_id, :city_id,
+    :amount, :color, :updated_at, :picture, :active, item_ids: [], other_item_ids: []
 
   filter :id, as: :numeric_range_filter
 
   filter :category_id, as: :search_select_filter,
-                       url: proc { current_admin_user.categories_url },
+                       url: "/admin/categories",
                        fields: [:name],
                        minimum_input_length: 0
 
-  filter :buyer_id, as: :search_select_filter,
-                    minimum_input_length: 0,
-                    url: '/admin/admin_users',
-                    fields: [:email],
-                    display_name: :email
+  filter :created_at, as: :date_time_picker_filter
 
   index do
     selectable_column
@@ -21,9 +17,11 @@ ActiveAdmin.register Invoice do
     tag_column :state, interactive: true
     bool_column :paid
     image_column :photo, style: :thumb
+    image_column :picture, style: :jpg_small
     attachment_column :attachment
     number_column :amount, as: :currency, unit: "$", separator: ","
     toggle_bool_column :active
+    column :created_at
     actions
   end
 
@@ -38,10 +36,15 @@ ActiveAdmin.register Invoice do
       list_row :details, localize: true
       image_row("Mi foto", :photo, style: :big, &:photo)
       attachment_row("My doc", :attachment, label: 'Download file', truncate: false, &:attachment)
+      image_row("Mi picture", :picture, image_options: { width: 100 }, &:picture)
       row :legal_date
       number_row("Monto", :amount, as: :human, &:amount)
       row :city
       bool_row :active
+    end
+
+    panel "Formtastic form" do
+      render partial: "formtastic_form"
     end
   end
 
@@ -59,7 +62,7 @@ ActiveAdmin.register Invoice do
       f.input :state
 
       f.input :category_id, as: :search_select,
-                            url: proc { current_admin_user.categories_url },
+                            url: proc { "/admin/categories" },
                             fields: [:name],
                             display_name: 'name',
                             minimum_input_length: 1,
@@ -82,7 +85,11 @@ ActiveAdmin.register Invoice do
 
       f.input :attachment
 
+      f.input :legal_date
+
       f.input :photo
+
+      f.input :picture, as: :file
 
       f.input :color, as: :color_picker,
                       palette: Invoice.colors
