@@ -6,8 +6,8 @@ class Invoice < ActiveRecord::Base
   extend ::Enumerize
   include Paperclip::Glue
   include AASM
+  include ImageUploader::Attachment(:picture)
 
-  belongs_to :buyer, class_name: "AdminUser", foreign_key: :client_id
   belongs_to :category
   belongs_to :city
   has_and_belongs_to_many :items
