@@ -2,7 +2,7 @@
 
 FactoryBot.define do
   factory :city do
-    name "MyString"
-    region nil
+    name { "MyString" }
+    region { nil }
   end
 end
