@@ -2,6 +2,6 @@
 
 FactoryBot.define do
   factory :country do
-    name "MyString"
+    name { "MyString" }
   end
 end
