@@ -2,7 +2,7 @@
 
 FactoryBot.define do
   factory :region do
-    name "MyString"
-    country nil
+    name { "MyString" }
+    country { nil }
   end
 end
