@@ -4,6 +4,7 @@ require 'spec_helper'
 require File.expand_path('dummy/config/environment', __dir__)
 require 'rspec/rails'
 require 'factory_bot_rails'
+require 'webdrivers'
 require 'capybara/rspec'
 require 'capybara/rails'
 require 'selenium-webdriver'
@@ -38,19 +39,23 @@ RSpec.configure do |config|
     DatabaseCleaner.clean
   end
 
+  # Cache the download of chrome driver for 1 day
+  Webdrivers.cache_time = 86_400
+
+  # Allow override of default path to Chrome (we use this in Travis)
+  Selenium::WebDriver::Chrome.path = ENV['CHROME_PATH'] if ENV['CHROME_PATH']
+
   Capybara.register_driver :chrome do |app|
     Capybara::Selenium::Driver.new(app, browser: :chrome)
   end
 
   Capybara.register_driver :headless_chrome do |app|
-    capabilities = Selenium::WebDriver::Remote::Capabilities.chrome(
-      chromeOptions: { args: %w(headless) }
-    )
-    Capybara::Selenium::Driver.new(app, browser: :chrome, desired_capabilities: capabilities)
+    args = ['no-sandbox', 'headless', 'disable-gpu', 'remote-debugging-port=9222']
+    options = Selenium::WebDriver::Chrome::Options.new(args: args)
+    Capybara::Selenium::Driver.new(app, browser: :chrome, options: options)
   end
 
-  # change to :chrome if you want to see the browser running.
-  Capybara.javascript_driver = :headless_chrome
+  Capybara.javascript_driver = ENV.fetch('JS_DRIVER', :headless_chrome).to_sym
 
   config.include ActiveAdminHelpers
   config.include DataLoaders
